server:start:watch restarts on project file save
