+++ server/nodemon.json
@@ -1,0 +1,6 @@
+{
+  "watch": ["server/**/*.ts"],
+  "ignore": ["server/**/*.spec.ts"],
+  "ext": "ts js json",
+  "exec": "ts-node --project ./server/tsconfig.json ./server/index.ts"
+}
